requirements.txt
asgiref==3.8.1
attrs==24.2.0
build==1.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.1.1
django-auditlog==3.0.0
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-import-export==4.1.1
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.15.4
idna==3.7
inflection==0.5.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupPy==1.14
MarkupSafe==2.1.5
numpy==2.1.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pudb==2024.1.2
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
urwid==2.6.15
urwid_readline==0.14
virtualenv==20.26.3
wcwidth==0.2.13
wheel==0.44.0
xlrd==2.0.1
xlwt==1.3.0
